toolsadded

+++ baseConverter.py
@@ -5,6 +5,6 @@ from tools.Base import *
 num1 = Base16()
 num2 = Base64("LrfAN")
 
-num1.setDecimal(num2.getDecimal())
+num1.setInt(num2.getInt())
 
 print(num1)
